refactor: extract ebus to root and rename to ebuse
Move all files from /ebus directory to repository root and rename
the project from ebus to ebuse throughout the codebase.

Changes:
- Moved all files from /ebus to repository root
- Renamed package from ebus to ebuse
- Updated import paths: github.com/jilio/ebus → github.com/jilio/ebuse
- Updated all documentation (README, PRODUCTION, MULTI-TENANT)
- Updated configuration files and examples
- Updated .gitignore to reflect new binary name
