+++ .gitignore
@@ -1,5 +1,5 @@
 # Binaries
-ebus
+ebuse
 *.exe
 *.exe~
 *.dll
