+++ config.go
@@ -1,4 +1,4 @@
-package ebus
+package ebuse
 
 import (
 	"os"
