+++ internal/store/sqlite_test.go
