+++ pkg/server/server.go
@@ -6,59 +6,73 @@ import (
 	"fmt"
 	"log"
 	"net/http"
+	"os"
 	"strconv"
 	"strings"
 	"time"
 
-	"github.com/jilio/ebus/internal/store"
+	"github.com/jilio/ebuse/internal/store"
 )
 
-// MultiTenantServer provides HTTP API with multi-tenant support
-type MultiTenantServer struct {
-	tenantManager TenantManager
-	mux           *http.ServeMux
-	rateLimiter   *rateLimiter
-	config        *Config
+// Server provides HTTP API for remote event storage
+type Server struct {
+	store       *store.SQLiteStore
+	apiKey      string
+	mux         *http.ServeMux
+	rateLimiter *rateLimiter
 }
 
-// TenantManager interface for managing multiple tenants
-type TenantManager interface {
-	GetStore(apiKey string) (*store.SQLiteStore, string, bool)
-	GetAllTenants() []string
-	Close() error
+// Config holds server configuration
+type Config struct {
+	RateLimit      int  // Requests per second per IP
+	RateBurst      int  // Burst size for rate limiter
+	EnableGzip     bool // Enable gzip compression
 }
 
-// NewMultiTenant creates a new multi-tenant server
-func NewMultiTenant(tenantManager TenantManager, config *Config) *MultiTenantServer {
-	if config == nil {
-		config = DefaultConfig()
+// DefaultConfig returns production-ready defaults
+func DefaultConfig() *Config {
+	return &Config{
+		RateLimit:  100, // 100 req/s per IP
+		RateBurst:  200, // Allow bursts up to 200
+		EnableGzip: true,
+	}
+}
+
+// New creates a new event storage server (deprecated: use NewWithConfig)
+func New(store *store.SQLiteStore) *Server {
+	apiKey := os.Getenv("API_KEY")
+	if apiKey == "" {
+		log.Fatal("API_KEY environment variable must be set")
+	}
+	return NewWithConfig(store, DefaultConfig(), apiKey)
+}
+
+// NewWithConfig creates a server with custom configuration
+func NewWithConfig(store *store.SQLiteStore, config *Config, apiKey string) *Server {
+	s := &Server{
+		store:       store,
+		apiKey:      apiKey,
+		mux:         http.NewServeMux(),
+		rateLimiter: newRateLimiter(config.RateLimit, config.RateBurst),
 	}
 
-	s := &MultiTenantServer{
-		tenantManager: tenantManager,
-		mux:           http.NewServeMux(),
-		rateLimiter:   newRateLimiter(config.RateLimit, config.RateBurst),
-		config:        config,
-	}
-
-	s.setupRoutes()
+	s.setupRoutes(config)
 	return s
 }
 
-func (s *MultiTenantServer) setupRoutes() {
+func (s *Server) setupRoutes(config *Config) {
 	// Apply middleware chain: rate limit -> auth -> compression -> handler
-	s.mux.HandleFunc("/events", s.chain(s.handleEvents, s.config.EnableGzip))
-	s.mux.HandleFunc("/events/batch", s.chain(s.handleBatchEvents, s.config.EnableGzip))
-	s.mux.HandleFunc("/events/stream", s.chain(s.handleStreamEvents, s.config.EnableGzip))
+	s.mux.HandleFunc("/events", s.chain(s.handleEvents, config.EnableGzip))
+	s.mux.HandleFunc("/events/batch", s.chain(s.handleBatchEvents, config.EnableGzip))
+	s.mux.HandleFunc("/events/stream", s.chain(s.handleStreamEvents, config.EnableGzip))
 	s.mux.HandleFunc("/position", s.chain(s.handlePosition, false))
 	s.mux.HandleFunc("/subscriptions/", s.chain(s.handleSubscriptions, false))
 	s.mux.HandleFunc("/health", s.handleHealth)
 	s.mux.HandleFunc("/metrics", s.authMiddleware(s.handleMetrics))
-	s.mux.HandleFunc("/tenants", s.authMiddleware(s.handleTenants))
 }
 
 // chain applies middleware in order: rate limit -> auth -> optional compression
-func (s *MultiTenantServer) chain(handler http.HandlerFunc, enableCompression bool) http.HandlerFunc {
+func (s *Server) chain(handler http.HandlerFunc, enableCompression bool) http.HandlerFunc {
 	h := handler
 	if enableCompression {
 		h = compressionMiddleware(h)
@@ -68,8 +82,8 @@ func (s *MultiTenantServer) chain(handler http.HandlerFunc, enableCompression bo
 	return h
 }
 
-// authMiddleware validates API key and injects tenant context
-func (s *MultiTenantServer) authMiddleware(next http.HandlerFunc) http.HandlerFunc {
+// authMiddleware validates the API_KEY header
+func (s *Server) authMiddleware(next http.HandlerFunc) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
 		apiKey := r.Header.Get("X-API-Key")
 		if apiKey == "" {
@@ -79,35 +93,17 @@ func (s *MultiTenantServer) authMiddleware(next http.HandlerFunc) http.HandlerFu
 			}
 		}
 
-		if apiKey == "" {
-			http.Error(w, "API key required", http.StatusUnauthorized)
+		if apiKey != s.apiKey {
+			http.Error(w, "Unauthorized", http.StatusUnauthorized)
 			return
 		}
 
-		// Get store for this API key
-		tenantStore, tenantName, ok := s.tenantManager.GetStore(apiKey)
-		if !ok {
-			http.Error(w, "Invalid API key", http.StatusUnauthorized)
-			return
-		}
-
-		// Inject tenant info into context
-		ctx := context.WithValue(r.Context(), "tenant_store", tenantStore)
-		ctx = context.WithValue(ctx, "tenant_name", tenantName)
-		next(w, r.WithContext(ctx))
+		next(w, r)
 	}
 }
 
-// getTenantStore extracts tenant store from context
-func getTenantStore(r *http.Request) (*store.SQLiteStore, string) {
-	tenantStore := r.Context().Value("tenant_store").(*store.SQLiteStore)
-	tenantName := r.Context().Value("tenant_name").(string)
-	return tenantStore, tenantName
-}
-
-// Event handlers (same as single-tenant but use tenant-specific store)
-
-func (s *MultiTenantServer) handleEvents(w http.ResponseWriter, r *http.Request) {
+// handleEvents handles both GET (load) and POST (save) for events
+func (s *Server) handleEvents(w http.ResponseWriter, r *http.Request) {
 	switch r.Method {
 	case http.MethodPost:
 		s.saveEvent(w, r)
@@ -118,9 +114,7 @@ func (s *MultiTenantServer) handleEvents(w http.ResponseWriter, r *http.Request)
 	}
 }
 
-func (s *MultiTenantServer) saveEvent(w http.ResponseWriter, r *http.Request) {
-	tenantStore, _ := getTenantStore(r)
-
+func (s *Server) saveEvent(w http.ResponseWriter, r *http.Request) {
 	var event store.StoredEvent
 	if err := json.NewDecoder(r.Body).Decode(&event); err != nil {
 		http.Error(w, fmt.Sprintf("Invalid request body: %v", err), http.StatusBadRequest)
@@ -130,7 +124,7 @@ func (s *MultiTenantServer) saveEvent(w http.ResponseWriter, r *http.Request) {
 	ctx, cancel := context.WithTimeout(r.Context(), 5*time.Second)
 	defer cancel()
 
-	if err := tenantStore.Save(ctx, &event); err != nil {
+	if err := s.store.Save(ctx, &event); err != nil {
 		http.Error(w, fmt.Sprintf("Failed to save event: %v", err), http.StatusInternalServerError)
 		return
 	}
@@ -139,9 +133,7 @@ func (s *MultiTenantServer) saveEvent(w http.ResponseWriter, r *http.Request) {
 	json.NewEncoder(w).Encode(event)
 }
 
-func (s *MultiTenantServer) loadEvents(w http.ResponseWriter, r *http.Request) {
-	tenantStore, _ := getTenantStore(r)
-
+func (s *Server) loadEvents(w http.ResponseWriter, r *http.Request) {
 	fromStr := r.URL.Query().Get("from")
 	toStr := r.URL.Query().Get("to")
 
@@ -163,7 +155,7 @@ func (s *MultiTenantServer) loadEvents(w http.ResponseWriter, r *http.Request) {
 	ctx, cancel := context.WithTimeout(r.Context(), 30*time.Second)
 	defer cancel()
 
-	events, err := tenantStore.Load(ctx, from, to)
+	events, err := s.store.Load(ctx, from, to)
 	if err != nil {
 		http.Error(w, fmt.Sprintf("Failed to load events: %v", err), http.StatusInternalServerError)
 		return
@@ -173,13 +165,12 @@ func (s *MultiTenantServer) loadEvents(w http.ResponseWriter, r *http.Request) {
 	json.NewEncoder(w).Encode(events)
 }
 
-func (s *MultiTenantServer) handleBatchEvents(w http.ResponseWriter, r *http.Request) {
+// handleBatchEvents handles batch event insertion
+func (s *Server) handleBatchEvents(w http.ResponseWriter, r *http.Request) {
 	if r.Method != http.MethodPost {
 		http.Error(w, "Method not allowed", http.StatusMethodNotAllowed)
 		return
 	}
-
-	tenantStore, _ := getTenantStore(r)
 
 	var events []*store.StoredEvent
 	if err := json.NewDecoder(r.Body).Decode(&events); err != nil {
@@ -195,26 +186,25 @@ func (s *MultiTenantServer) handleBatchEvents(w http.ResponseWriter, r *http.Req
 	ctx, cancel := context.WithTimeout(r.Context(), 30*time.Second)
 	defer cancel()
 
-	if err := tenantStore.SaveBatch(ctx, events); err != nil {
+	if err := s.store.SaveBatch(ctx, events); err != nil {
 		http.Error(w, fmt.Sprintf("Failed to save batch: %v", err), http.StatusInternalServerError)
 		return
 	}
 
 	w.Header().Set("Content-Type", "application/json")
 	json.NewEncoder(w).Encode(map[string]interface{}{
-		"saved":          len(events),
+		"saved": len(events),
 		"first_position": events[0].Position,
-		"last_position":  events[len(events)-1].Position,
+		"last_position": events[len(events)-1].Position,
 	})
 }
 
-func (s *MultiTenantServer) handleStreamEvents(w http.ResponseWriter, r *http.Request) {
+// handleStreamEvents streams events for large replays
+func (s *Server) handleStreamEvents(w http.ResponseWriter, r *http.Request) {
 	if r.Method != http.MethodGet {
 		http.Error(w, "Method not allowed", http.StatusMethodNotAllowed)
 		return
 	}
-
-	tenantStore, _ := getTenantStore(r)
 
 	fromStr := r.URL.Query().Get("from")
 	batchSizeStr := r.URL.Query().Get("batch_size")
@@ -235,13 +225,15 @@ func (s *MultiTenantServer) handleStreamEvents(w http.ResponseWriter, r *http.Re
 
 	ctx := r.Context()
 
+	// Set headers for streaming
 	w.Header().Set("Content-Type", "application/json")
 	w.Header().Set("Transfer-Encoding", "chunked")
 
+	// Use JSON array streaming
 	w.Write([]byte("["))
 	first := true
 
-	err = tenantStore.LoadStream(ctx, from, batchSize, func(batch []*store.StoredEvent) error {
+	err = s.store.LoadStream(ctx, from, batchSize, func(batch []*store.StoredEvent) error {
 		for _, event := range batch {
 			if !first {
 				w.Write([]byte(","))
@@ -254,6 +246,7 @@ func (s *MultiTenantServer) handleStreamEvents(w http.ResponseWriter, r *http.Re
 			}
 			w.Write(data)
 
+			// Flush to client
 			if flusher, ok := w.(http.Flusher); ok {
 				flusher.Flush()
 			}
@@ -268,18 +261,16 @@ func (s *MultiTenantServer) handleStreamEvents(w http.ResponseWriter, r *http.Re
 	w.Write([]byte("]"))
 }
 
-func (s *MultiTenantServer) handlePosition(w http.ResponseWriter, r *http.Request) {
+func (s *Server) handlePosition(w http.ResponseWriter, r *http.Request) {
 	if r.Method != http.MethodGet {
 		http.Error(w, "Method not allowed", http.StatusMethodNotAllowed)
 		return
 	}
 
-	tenantStore, _ := getTenantStore(r)
-
 	ctx, cancel := context.WithTimeout(r.Context(), 5*time.Second)
 	defer cancel()
 
-	position, err := tenantStore.GetPosition(ctx)
+	position, err := s.store.GetPosition(ctx)
 	if err != nil {
 		http.Error(w, fmt.Sprintf("Failed to get position: %v", err), http.StatusInternalServerError)
 		return
@@ -289,9 +280,8 @@ func (s *MultiTenantServer) handlePosition(w http.ResponseWriter, r *http.Reques
 	json.NewEncoder(w).Encode(map[string]int64{"position": position})
 }
 
-func (s *MultiTenantServer) handleSubscriptions(w http.ResponseWriter, r *http.Request) {
-	tenantStore, _ := getTenantStore(r)
-
+func (s *Server) handleSubscriptions(w http.ResponseWriter, r *http.Request) {
+	// Extract subscription ID from path: /subscriptions/{id}/position
 	path := strings.TrimPrefix(r.URL.Path, "/subscriptions/")
 	parts := strings.Split(path, "/")
 
@@ -304,15 +294,15 @@ func (s *MultiTenantServer) handleSubscriptions(w http.ResponseWriter, r *http.R
 
 	switch r.Method {
 	case http.MethodPost, http.MethodPut:
-		s.saveSubscriptionPosition(w, r, tenantStore, subscriptionID)
+		s.saveSubscriptionPosition(w, r, subscriptionID)
 	case http.MethodGet:
-		s.loadSubscriptionPosition(w, r, tenantStore, subscriptionID)
+		s.loadSubscriptionPosition(w, r, subscriptionID)
 	default:
 		http.Error(w, "Method not allowed", http.StatusMethodNotAllowed)
 	}
 }
 
-func (s *MultiTenantServer) saveSubscriptionPosition(w http.ResponseWriter, r *http.Request, tenantStore *store.SQLiteStore, subscriptionID string) {
+func (s *Server) saveSubscriptionPosition(w http.ResponseWriter, r *http.Request, subscriptionID string) {
 	var req struct {
 		Position int64 `json:"position"`
 	}
@@ -324,7 +314,7 @@ func (s *MultiTenantServer) saveSubscriptionPosition(w http.ResponseWriter, r *h
 	ctx, cancel := context.WithTimeout(r.Context(), 5*time.Second)
 	defer cancel()
 
-	if err := tenantStore.SaveSubscriptionPosition(ctx, subscriptionID, req.Position); err != nil {
+	if err := s.store.SaveSubscriptionPosition(ctx, subscriptionID, req.Position); err != nil {
 		http.Error(w, fmt.Sprintf("Failed to save subscription position: %v", err), http.StatusInternalServerError)
 		return
 	}
@@ -332,11 +322,11 @@ func (s *MultiTenantServer) saveSubscriptionPosition(w http.ResponseWriter, r *h
 	w.WriteHeader(http.StatusNoContent)
 }
 
-func (s *MultiTenantServer) loadSubscriptionPosition(w http.ResponseWriter, r *http.Request, tenantStore *store.SQLiteStore, subscriptionID string) {
+func (s *Server) loadSubscriptionPosition(w http.ResponseWriter, r *http.Request, subscriptionID string) {
 	ctx, cancel := context.WithTimeout(r.Context(), 5*time.Second)
 	defer cancel()
 
-	position, err := tenantStore.LoadSubscriptionPosition(ctx, subscriptionID)
+	position, err := s.store.LoadSubscriptionPosition(ctx, subscriptionID)
 	if err != nil {
 		http.Error(w, fmt.Sprintf("Failed to load subscription position: %v", err), http.StatusInternalServerError)
 		return
@@ -346,46 +336,51 @@ func (s *MultiTenantServer) loadSubscriptionPosition(w http.ResponseWriter, r *h
 	json.NewEncoder(w).Encode(map[string]int64{"position": position})
 }
 
-func (s *MultiTenantServer) handleHealth(w http.ResponseWriter, r *http.Request) {
+// handleHealth provides health check endpoint
+func (s *Server) handleHealth(w http.ResponseWriter, r *http.Request) {
+	ctx, cancel := context.WithTimeout(r.Context(), 2*time.Second)
+	defer cancel()
+
+	// Check database connectivity
+	_, err := s.store.GetPosition(ctx)
+	if err != nil {
+		w.WriteHeader(http.StatusServiceUnavailable)
+		json.NewEncoder(w).Encode(map[string]string{
+			"status": "unhealthy",
+			"error":  err.Error(),
+		})
+		return
+	}
+
 	w.Header().Set("Content-Type", "application/json")
 	json.NewEncoder(w).Encode(map[string]string{
 		"status": "healthy",
 	})
 }
 
-func (s *MultiTenantServer) handleMetrics(w http.ResponseWriter, r *http.Request) {
-	tenantStore, tenantName := getTenantStore(r)
-
+// handleMetrics provides basic metrics
+func (s *Server) handleMetrics(w http.ResponseWriter, r *http.Request) {
 	ctx, cancel := context.WithTimeout(r.Context(), 5*time.Second)
 	defer cancel()
 
-	position, _ := tenantStore.GetPosition(ctx)
+	position, _ := s.store.GetPosition(ctx)
 
 	w.Header().Set("Content-Type", "application/json")
 	json.NewEncoder(w).Encode(map[string]interface{}{
-		"tenant":       tenantName,
 		"total_events": position,
 		"timestamp":    time.Now().Unix(),
 	})
 }
 
-func (s *MultiTenantServer) handleTenants(w http.ResponseWriter, r *http.Request) {
-	tenants := s.tenantManager.GetAllTenants()
-
-	w.Header().Set("Content-Type", "application/json")
-	json.NewEncoder(w).Encode(map[string]interface{}{
-		"tenants": tenants,
-		"count":   len(tenants),
-	})
-}
-
-func (s *MultiTenantServer) Close() error {
+// Close stops the server and cleans up resources
+func (s *Server) Close() error {
 	if s.rateLimiter != nil {
 		s.rateLimiter.Stop()
 	}
-	return s.tenantManager.Close()
+	return nil
 }
 
-func (s *MultiTenantServer) ServeHTTP(w http.ResponseWriter, r *http.Request) {
+// ServeHTTP implements http.Handler
+func (s *Server) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	s.mux.ServeHTTP(w, r)
 }
